- debugging of the convolution layer - refactoring

+++ src/snOperator/src/Operator/fullyConnected.cpp
@@ -134,7 +134,7 @@ std::vector<std::string> FullyConnected::Do(const learningParam& lernPrm, const 
 
 			inFwTns_ = *neighbOpr[0]->getOutput();
 
-			int sz = neighbOpr.size();
+			size_t sz = neighbOpr.size();
 			for (size_t i = 1; i < sz; ++i)
 				inFwTns_ += *neighbOpr[i]->getOutput();
 
@@ -144,7 +144,7 @@ std::vector<std::string> FullyConnected::Do(const learningParam& lernPrm, const 
 
 			inBwTns_ = *neighbOpr[0]->getGradient();
 
-			int sz = neighbOpr.size();
+			size_t sz = neighbOpr.size();
 			for (size_t i = 1; i < sz; ++i)
 				inBwTns_ += *neighbOpr[i]->getGradient();
